requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24.0
django-cors-headers>=4.3
psycopg[binary]>=3.1
requests>=2.31
python-dotenv>=1.0
